Move Type impl where its defined
Also add Type::{non_blocking, cloexec} on OSes that support it.
