+++ src/sys/unix.rs
@@ -24,13 +24,16 @@ use std::time::{Duration, Instant};
 
 use libc::{self, c_void, socklen_t, ssize_t};
 
-use crate::Domain;
+use crate::{Domain, Type};
 
 #[allow(non_camel_case_types)]
 pub(crate) type c_int = libc::c_int;
 
 // Used in `Domain`.
 pub(crate) use libc::{AF_INET, AF_INET6};
+// Used in `Type`.
+pub(crate) use libc::{SOCK_DGRAM, SOCK_RAW, SOCK_SEQPACKET, SOCK_STREAM};
+
 cfg_if::cfg_if! {
     if #[cfg(any(target_os = "dragonfly", target_os = "freebsd",
                  target_os = "ios", target_os = "macos",
@@ -42,19 +45,6 @@ cfg_if::cfg_if! {
     } else {
         use libc::IPV6_ADD_MEMBERSHIP;
         use libc::IPV6_DROP_MEMBERSHIP;
-    }
-}
-// Used in `Type`.
-pub(crate) use libc::{SOCK_RAW, SOCK_SEQPACKET};
-
-cfg_if::cfg_if! {
-    if #[cfg(any(target_os = "linux", target_os = "android",
-                 target_os = "dragonfly", target_os = "freebsd",
-                 target_os = "openbsd", target_os = "netbsd",
-                 target_os = "haiku", target_os = "bitrig"))] {
-        use libc::MSG_NOSIGNAL;
-    } else {
-        const MSG_NOSIGNAL: c_int = 0x0;
     }
 }
 
@@ -91,6 +81,45 @@ impl Domain {
     #[cfg(target_os = "linux")]
     pub fn packet() -> Domain {
         Domain(libc::AF_PACKET)
+    }
+}
+
+/// Unix only API.
+impl Type {
+    /// Set `SOCK_NONBLOCK` on the `Type`.
+    ///
+    /// # Notes
+    ///
+    /// This function is only available on Android, DragonFlyBSD, FreeBSD,
+    /// Linux, NetBSD and OpenBSD.
+    #[cfg(any(
+        target_os = "android",
+        target_os = "dragonfly",
+        target_os = "freebsd",
+        target_os = "linux",
+        target_os = "netbsd",
+        target_os = "openbsd"
+    ))]
+    pub fn non_blocking(self) -> Type {
+        Type(self.0 | libc::SOCK_NONBLOCK)
+    }
+
+    /// Set `SOCK_CLOEXEC` on the `Type`.
+    ///
+    /// # Notes
+    ///
+    /// This function is only available on Android, DragonFlyBSD, FreeBSD,
+    /// Linux, NetBSD and OpenBSD.
+    #[cfg(any(
+        target_os = "android",
+        target_os = "dragonfly",
+        target_os = "freebsd",
+        target_os = "linux",
+        target_os = "netbsd",
+        target_os = "openbsd"
+    ))]
+    pub fn cloexec(self) -> Type {
+        Type(self.0 | libc::SOCK_CLOEXEC)
     }
 }
